rdrf #1887 decorator modified
