+++ rdrf/rdrf/models/definition/models.py
@@ -20,7 +20,8 @@ from django.forms.models import model_to_dict
 from django.utils.safestring import mark_safe
 from django.core.exceptions import PermissionDenied
 
-from rdrf.helpers.utils import format_date, parse_iso_datetime, contains_blacklisted_words, use_cache
+from rdrf.helpers.decorators import use_cache
+from rdrf.helpers.utils import format_date, parse_iso_datetime, contains_blacklisted_words
 from rdrf.helpers.utils import LinkWrapper
 from rdrf.events.events import EventType
 
@@ -222,7 +223,7 @@ class Registry(models.Model):
         return self.registry_type == RegistryType.NORMAL
 
     @property
-    @use_cache(key="registry_metadata")
+    @use_cache
     def metadata(self):
         if self.metadata_json:
             try:
